investigate #295 — TIOCNOTTY didn't help; accept macOS pty quirk
Round 2 follow-up run on rocinante:

1. Build regression in 3e0b95d on macOS — libc::TIOCNOTTY is c_uint on
   Apple, c_ulong on Linux. Fixed with `as libc::c_ulong` cast at the
   call site; portable across all supported targets.

2. With the build fix in, test still fails TimedOut after SIGHUP in 8s
   (vs. the 3+h pre-bounded-reap baseline). Trace shows
   `TIOCNOTTY done — falling off main` at 0.391s as the LAST line,
   so the ioctl ran but the wedge is past `fn main`'s closing brace —
   in the Rust/libc exit epilogue.

   Hypothesis: kernel close-path on the slave pty waits for the master
   to drain. The harness's expect_exit_within polls waitpid without
   draining the master, so the slave's output buffer never empties,
   kernel revoke wedges, process sits in state ?NEs. TIOCNOTTY only
   changes session-leader status; it doesn't drain the buffer.

3. Per the doc's §3b decision table row 2: accept the kernel-side
   wedge. Keep cfg(target_os = "linux") on pty_smoke.rs +
   pty_paint_no_absolute.rs. Bounded reap (01399ba) is the durable
   shield. TIOCNOTTY stays in main as a defensive no-op for
   Linux/non-pty stdin.

   #295 should close as wontfix / accepted limitation. A real macOS
   fix would need master-side drain in the harness, which is not in
   scope here.

+++ crates/rush-cli/src/main.rs
@@ -272,7 +272,7 @@ fn main() {
     // observable effect is the explicit detach.
     #[cfg(unix)]
     unsafe {
-        libc::ioctl(libc::STDIN_FILENO, libc::TIOCNOTTY);
+        libc::ioctl(libc::STDIN_FILENO, libc::TIOCNOTTY as libc::c_ulong);
     }
     rush_line::trace!("main", "TIOCNOTTY done — falling off main");
 }
